Remove RecursiveDeleter and inline it in the one place it is used
